requirements.txt
ansicon==1.89.0
anyio==3.6.2
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
console==0.9907
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
ezenv==0.92
fastapi==0.95.0
Flask==2.2.3
fonttools==4.39.2
h11==0.14.0
idna==3.4
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jinxed==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mysql-connector-python==8.0.32
numpy==1.24.2
out==0.78
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
protobuf==3.20.3
pure-eval==0.2.2
pydantic==1.10.7
Pygments==2.15.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.2
requests==2.28.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.26.1
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
wcwidth==0.2.6
Werkzeug==2.2.3
